EP-2461 various code review followups
- Issue #12 avoid new code in __init__.py
- some TODO notes and tests related to ServiceRegistry
- initial implementation of cleaning up in Zookeeper and Traefic
